begin reorganizing tests to put directory-related tests in their own file
