+++ test/test_littlefs.c
@@ -1,51 +1,14 @@
 //#define LOG_LOCAL_LEVEL 4
-
-#include "esp_littlefs.h"
-#include "sdkconfig.h"
-
-#include <stdio.h>
-#include <fcntl.h>
-#include <stdlib.h>
-#include <string.h>
-#include <time.h>
-#include <sys/time.h>
-#include <sys/unistd.h>
-#include <sys/param.h>
-#include "unity.h"
-#include "test_utils.h"
-#include "esp_log.h"
-#include "esp_system.h"
-#include "esp_vfs.h"
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-#include "freertos/queue.h"
-#include "freertos/semphr.h"
-#include "esp_heap_caps.h"
-#include "esp_partition.h"
-#include "errno.h"
-#include "esp_idf_version.h"
-
-#if ESP_IDF_VERSION < ESP_IDF_VERSION_VAL(4, 3, 0)
-#define esp_rom_printf ets_printf
-#else
-#include "esp_rom_sys.h"
-#endif
+#include "test_littlefs_common.h"
 
 
-static const char littlefs_test_partition_label[] = "flash_test";
-static const char littlefs_test_hello_str[] = "Hello, World!\n";
-#define littlefs_base_path "/littlefs"
+
 
 static void test_littlefs_write_file_with_offset(const char *filename);
 static void test_littlefs_read_file_with_offset(const char *filename);
-static void test_littlefs_create_file_with_text(const char* name, const char* text);
 static void test_littlefs_overwrite_append(const char* filename);
-static void test_littlefs_read_file(const char* filename);
-static void test_littlefs_readdir_many_files(const char* dir_prefix);
 static void test_littlefs_open_max_files(const char* filename_prefix, size_t files_count);
 static void test_littlefs_concurrent_rw(const char* filename_prefix);
-static void test_setup();
-static void test_teardown();
 
 TEST_CASE("can initialize LittleFS in erased partition", "[littlefs]")
 {
@@ -407,147 +370,6 @@ TEST_CASE("can opendir root directory of FS", "[littlefs]")
     test_teardown();
 }
 
-TEST_CASE("mkdir, rmdir", "[littlefs]")
-{
-    test_setup();
-    const char filename_prefix[] = littlefs_base_path "/";
-
-    char name_dir1[64];
-    char name_dir2[64];
-    char name_dir2_file[64];
-
-    snprintf(name_dir1, sizeof(name_dir1), "%s1", filename_prefix);
-    snprintf(name_dir2, sizeof(name_dir2), "%s2", filename_prefix);
-    snprintf(name_dir2_file, sizeof(name_dir2_file), "%s2/1.txt", filename_prefix);
-
-    TEST_ASSERT_EQUAL(0, mkdir(name_dir1, 0755));
-    struct stat st;
-    TEST_ASSERT_EQUAL(0, stat(name_dir1, &st));
-    TEST_ASSERT_TRUE(st.st_mode & S_IFDIR);
-    TEST_ASSERT_FALSE(st.st_mode & S_IFREG);
-    TEST_ASSERT_EQUAL(0, rmdir(name_dir1));
-
-    // Attempt to stat a removed directory
-    TEST_ASSERT_EQUAL(-1, stat(name_dir1, &st));
-    TEST_ASSERT_EQUAL(ENOENT, errno);
-
-    TEST_ASSERT_EQUAL(0, mkdir(name_dir2, 0755));
-    test_littlefs_create_file_with_text(name_dir2_file, "foo\n");
-    TEST_ASSERT_EQUAL(0, stat(name_dir2, &st));
-    TEST_ASSERT_TRUE(st.st_mode & S_IFDIR);
-    TEST_ASSERT_FALSE(st.st_mode & S_IFREG);
-    TEST_ASSERT_EQUAL(0, stat(name_dir2_file, &st));
-    TEST_ASSERT_FALSE(st.st_mode & S_IFDIR);
-    TEST_ASSERT_TRUE(st.st_mode & S_IFREG);
-
-    // Can't remove directory, its not empty
-    TEST_ASSERT_EQUAL(-1, rmdir(name_dir2));
-    TEST_ASSERT_EQUAL(ENOTEMPTY, errno);
-
-    TEST_ASSERT_EQUAL(0, unlink(name_dir2_file));
-#if !CONFIG_LITTLEFS_SPIFFS_COMPAT
-    /* this will have already been deleted */
-    TEST_ASSERT_EQUAL(0, rmdir(name_dir2));
-#endif
-
-    test_teardown();
-}
-
-TEST_CASE("opendir, readdir, rewinddir, seekdir work as expected", "[littlefs]")
-{
-    test_setup();
-    const char dir_prefix[] = littlefs_base_path "/dir";
-
-    char name_dir_inner_file[64];
-    char name_dir_inner[64];
-    char name_dir_file3[64];
-    char name_dir_file2[64];
-    char name_dir_file1[64];
-
-    snprintf(name_dir_inner_file, sizeof(name_dir_inner_file), "%s/inner/3.txt", dir_prefix);
-    snprintf(name_dir_inner,      sizeof(name_dir_inner),      "%s/inner",       dir_prefix);
-    snprintf(name_dir_file3,      sizeof(name_dir_file2),      "%s/boo.bin",     dir_prefix);
-    snprintf(name_dir_file2,      sizeof(name_dir_file2),      "%s/2.txt",       dir_prefix);
-    snprintf(name_dir_file1,      sizeof(name_dir_file1),      "%s/1.txt",       dir_prefix);
-
-    /* Remove files/dirs that may exist */
-    unlink(name_dir_inner_file);
-    rmdir(name_dir_inner);
-    unlink(name_dir_file1);
-    unlink(name_dir_file2);
-    unlink(name_dir_file3);
-    rmdir(dir_prefix);
-
-    /* Create the files */
-    TEST_ASSERT_EQUAL(0, mkdir(dir_prefix, 0755));
-    TEST_ASSERT_EQUAL(0, mkdir(name_dir_inner, 0755));
-    test_littlefs_create_file_with_text(name_dir_file1, "1\n");
-    test_littlefs_create_file_with_text(name_dir_file2, "2\n");
-    test_littlefs_create_file_with_text(name_dir_file3, "\01\02\03");
-    test_littlefs_create_file_with_text(name_dir_inner_file, "3\n");
-
-    DIR* dir = opendir(dir_prefix);
-    TEST_ASSERT_NOT_NULL(dir);
-    int count = 0;
-    const char* names[4];
-    while( true ) {
-        struct dirent* de = readdir(dir);
-        if (!de) {
-            break;
-        }
-        if (strcasecmp(de->d_name, "1.txt") == 0) {
-            TEST_ASSERT_TRUE(de->d_type == DT_REG);
-            names[count] = "1.txt";
-            ++count;
-        } else if (strcasecmp(de->d_name, "2.txt") == 0) {
-            TEST_ASSERT_TRUE(de->d_type == DT_REG);
-            names[count] = "2.txt";
-            ++count;
-        } else if (strcasecmp(de->d_name, "inner") == 0) {
-            TEST_ASSERT_TRUE(de->d_type == DT_DIR);
-            names[count] = "inner";
-            ++count;
-        } else if (strcasecmp(de->d_name, "boo.bin") == 0) {
-            TEST_ASSERT_TRUE(de->d_type == DT_REG);
-            names[count] = "boo.bin";
-            ++count;
-        } else {
-            char buf[512] = { 0 };
-            snprintf(buf, sizeof(buf), "unexpected directory entry \"%s\"", de->d_name);
-            TEST_FAIL_MESSAGE(buf);
-        }
-    }
-    TEST_ASSERT_EQUAL(count, 4);
-
-    rewinddir(dir);
-    struct dirent* de = readdir(dir);
-    TEST_ASSERT_NOT_NULL(de);
-    TEST_ASSERT_EQUAL(0, strcasecmp(de->d_name, names[0]));
-    seekdir(dir, 3);
-    de = readdir(dir);
-    TEST_ASSERT_NOT_NULL(de);
-    TEST_ASSERT_EQUAL(0, strcasecmp(de->d_name, names[3]));
-    seekdir(dir, 1);
-    de = readdir(dir);
-    TEST_ASSERT_NOT_NULL(de);
-    TEST_ASSERT_EQUAL(0, strcasecmp(de->d_name, names[1]));
-    seekdir(dir, 2);
-    de = readdir(dir);
-    TEST_ASSERT_NOT_NULL(de);
-    TEST_ASSERT_EQUAL(0, strcasecmp(de->d_name, names[2]));
-
-    TEST_ASSERT_EQUAL(0, closedir(dir));
-
-    test_teardown();
-}
-
-TEST_CASE("readdir with large number of files", "[littlefs][timeout=30]")
-{
-    test_setup();
-    test_littlefs_readdir_many_files(littlefs_base_path "/dir2");
-    test_teardown();
-}
-
 TEST_CASE("multiple tasks can use same volume", "[littlefs]")
 {
     test_setup();
@@ -737,15 +559,6 @@ static void test_littlefs_read_file_with_offset(const char *filename)
     TEST_ASSERT_EQUAL(0, close(fd));
 }
 
-static void test_littlefs_create_file_with_text(const char* name, const char* text)
-{
-    printf("Writing to \"%s\"\n", name);
-    FILE* f = fopen(name, "wb");
-    TEST_ASSERT_NOT_NULL(f);
-    TEST_ASSERT_TRUE(fputs(text, f) != EOF);
-    TEST_ASSERT_EQUAL(0, fclose(f));
-}
-
 static void test_littlefs_overwrite_append(const char* filename)
 {
     /* Create new file with 'aaaa' */
@@ -779,78 +592,6 @@ static void test_littlefs_overwrite_append(const char* filename)
     TEST_ASSERT_EQUAL(4, fread(buf, 1, 8, f_r)); // trying to read 8 bytes, only expecting 4
     TEST_ASSERT_EQUAL_STRING_LEN("cccc", buf, 4);
     TEST_ASSERT_EQUAL(0, fclose(f_r));
-}
-
-static void test_littlefs_read_file(const char* filename)
-{
-    FILE* f = fopen(filename, "r");
-    TEST_ASSERT_NOT_NULL(f);
-    char buf[32] = { 0 };
-    int cb = fread(buf, 1, sizeof(buf), f);
-    TEST_ASSERT_EQUAL(strlen(littlefs_test_hello_str), cb);
-    TEST_ASSERT_EQUAL(0, strcmp(littlefs_test_hello_str, buf));
-    TEST_ASSERT_EQUAL(0, fclose(f));
-}
-
-static void test_littlefs_readdir_many_files(const char* dir_prefix)
-{
-    const int n_files = 40;
-    const int n_folders = 4;
-    unsigned char file_count[n_files * n_folders];
-    memset(file_count, 0, sizeof(file_count)/sizeof(file_count[0]));
-    char file_name[ESP_VFS_PATH_MAX + CONFIG_LITTLEFS_OBJ_NAME_LEN];
-
-    /* clean stale files before the test */
-    mkdir(dir_prefix, 0755);
-    DIR* dir = opendir(dir_prefix);
-    if (dir) {
-        while (true) {
-            struct dirent* de = readdir(dir);
-            if (!de) {
-                break;
-            }
-            int len = snprintf(file_name, sizeof(file_name), "%s/%s", dir_prefix, de->d_name);
-            assert(len < sizeof(file_name));
-            unlink(file_name);
-        }
-    }
-
-    /* create files */
-    for (int d = 0; d < n_folders; ++d) {
-        printf("filling directory %d\n", d);
-        snprintf(file_name, sizeof(file_name), "%s/%d", dir_prefix, d);
-        mkdir( file_name, 0755 );
-        for (int f = 0; f < n_files; ++f) {
-            snprintf(file_name, sizeof(file_name), "%s/%d/%d.txt", dir_prefix, d, f);
-            test_littlefs_create_file_with_text(file_name, file_name);
-        }
-    }
-
-    /* list files */
-    for (int d = 0; d < n_folders; ++d) {
-        printf("listing files in directory %d\n", d);
-        snprintf(file_name, sizeof(file_name), "%s/%d", dir_prefix, d);
-        dir = opendir(file_name);
-        TEST_ASSERT_NOT_NULL(dir);
-        while (true) {
-            struct dirent* de = readdir(dir);
-            if (!de) {
-                break;
-            }
-            int file_id;
-            TEST_ASSERT_EQUAL(1, sscanf(de->d_name, "%d.txt", &file_id));
-            file_count[file_id + d * n_files]++;
-        }
-        closedir(dir);
-    }
-
-    /* check that all created files have been seen */
-    for (int d = 0; d < n_folders; ++d) {
-        printf("checking that all files have been found in directory %d\n", d);
-        for (int f = 0; f < n_files; ++f) {
-            TEST_ASSERT_EQUAL(1, file_count[f + d * n_files]);
-        }
-    }
 }
 
 static void test_littlefs_open_max_files(const char* filename_prefix, size_t files_count)
@@ -1272,22 +1013,4 @@ TEST_CASE("fcntl get flags", "[littlefs]")
     TEST_ASSERT_EQUAL(0, ret);
 
     test_teardown();
-}
-
-static void test_setup() {
-    esp_littlefs_format(littlefs_test_partition_label);
-    const esp_vfs_littlefs_conf_t conf = {
-        .base_path = littlefs_base_path,
-        .partition_label = littlefs_test_partition_label,
-        .format_if_mount_failed = true
-    };
-    TEST_ESP_OK(esp_vfs_littlefs_register(&conf));
-    TEST_ASSERT_TRUE( heap_caps_check_integrity_all(true) );
-    printf("Test setup complete.\n");
-}
-
-static void test_teardown(){
-    TEST_ESP_OK(esp_vfs_littlefs_unregister(littlefs_test_partition_label));
-    TEST_ASSERT_TRUE( heap_caps_check_integrity_all(true) );
-    printf("Test teardown complete.\n");
 }
